+++ src/isomedia/tx3g.c
@@ -371,28 +371,28 @@ u32 gf_isom_text_sample_size(GF_TextSample *samp)
 	size = 2 + samp->len;
 	if (samp->styles) {
 		gf_isom_box_size((GF_Box *)samp->styles);
-		size += samp->styles->size;
+		size += (u32) samp->styles->size;
 	}
 	if (samp->highlight_color) {
 		gf_isom_box_size((GF_Box *)samp->highlight_color);
-		size += samp->highlight_color->size;
+		size += (u32) samp->highlight_color->size;
 	}
 	if (samp->scroll_delay) {
 		gf_isom_box_size((GF_Box *)samp->scroll_delay);
-		size += samp->scroll_delay->size;
+		size += (u32) samp->scroll_delay->size;
 	}
 	if (samp->box) {
 		gf_isom_box_size((GF_Box *)samp->box);
-		size += samp->box->size;
+		size += (u32) samp->box->size;
 	}
 	if (samp->wrap) {
 		gf_isom_box_size((GF_Box *)samp->wrap);
-		size += samp->wrap->size;
+		size += (u32) samp->wrap->size;
 	}
 	i=0;
 	while ((a = (GF_Box*)gf_list_enum(samp->others, &i))) {
 		gf_isom_box_size((GF_Box *)a);
-		size += a->size;
+		size += (u32) a->size;
 	}
 	return size;
 }
